requirements.txt
amqp==2.6.1
-e git+https://github.com/ujjawal-shrivastava/ariadne.git@fdb8640864d16ffaa9f314db087889f2065b2a4a#egg=ariadne
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.7
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
graphql-core==3.0.5
isort==4.3.21
itsdangerous==1.1.0
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
starlette==0.13.8
toml==0.10.1
typing-extensions==3.7.4.3
vine==1.3.0
whitenoise==5.2.0
wrapt==1.12.1
